pyproject.toml
[tool.poetry]
name = "testovoe-steelsmart"
version = "0.1.0"
description = ""
authors = ["Zaks Danil <zaksdanil1913@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
django-rest-swagger = "^2.2.0"
black = "^24.4.2"
pillow = "^10.4.0"
django-ckeditor = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
